Fixed Recursion Issue - Changed Load Function
There was an issue with recursion depth when saving using Pickle. Added
a line in the RNN.py code to increase the allowed recursion depth and it
ran. Seems to have some small issue bringing in the data, not sure where
that issue is those. Check "GNET.txt" and "GNET2.txt" to view the
differences.

Load Function now returns a model, rather than having it take in an
uninitialized variable.